update the deployment ssm parameters
