+++ infrastructure/lib/app-stack.ts
@@ -3,7 +3,6 @@ import * as cdk from "aws-cdk-lib";
 import * as ecs from "aws-cdk-lib/aws-ecs";
 import * as ec2 from "aws-cdk-lib/aws-ec2";
 import * as efs from "aws-cdk-lib/aws-efs";
-import * as ecr from "aws-cdk-lib/aws-ecr";
 import * as iam from "aws-cdk-lib/aws-iam";
 import * as logs from "aws-cdk-lib/aws-logs";
 import * as elbv2 from "aws-cdk-lib/aws-elasticloadbalancingv2";
@@ -405,400 +404,5 @@ export class cprositeStack extends cdk.Stack {
         stringValue: value,
       });
     }
-  }
-}
-
-interface NetworkStackProps extends cdk.StackProps {
-  env?: cdk.Environment;
-  tier: string;
-  existingVpcId?: string;
-  existingAlbSecurityGroupId?: string;
-  existingEcsSecurityGroupId?: string;
-}
-
-// Compatibility stack: supports either importing existing networking or creating a default VPC layout.
-export class NetworkStack extends cdk.Stack {
-  public readonly vpc: ec2.IVpc;
-  public readonly albSecurityGroup: ec2.ISecurityGroup;
-  public readonly ecsSecurityGroup: ec2.ISecurityGroup;
-
-  constructor(scope: Construct, id: string, props: NetworkStackProps) {
-    super(scope, id, props);
-
-    const hasAnyExistingNetworkId =
-      !!props.existingVpcId ||
-      !!props.existingAlbSecurityGroupId ||
-      !!props.existingEcsSecurityGroupId;
-    const hasAllExistingNetworkIds =
-      !!props.existingVpcId &&
-      !!props.existingAlbSecurityGroupId &&
-      !!props.existingEcsSecurityGroupId;
-
-    if (hasAnyExistingNetworkId && !hasAllExistingNetworkIds) {
-      throw new Error(
-        "When using existing network, set EXISTING_VPC_ID, EXISTING_ALB_SG_ID, and EXISTING_ECS_SG_ID together."
-      );
-    }
-
-    if (hasAllExistingNetworkIds) {
-      // Reuse an existing network footprint managed outside this stack.
-      this.vpc = ec2.Vpc.fromLookup(this, "ExistingVpc", {
-        vpcId: props.existingVpcId,
-      });
-      this.albSecurityGroup = ec2.SecurityGroup.fromSecurityGroupId(
-        this,
-        "ExistingAlbSecurityGroup",
-        props.existingAlbSecurityGroupId,
-        { mutable: false }
-      );
-      this.ecsSecurityGroup = ec2.SecurityGroup.fromSecurityGroupId(
-        this,
-        "ExistingEcsSecurityGroup",
-        props.existingEcsSecurityGroupId,
-        { mutable: false }
-      );
-    } else {
-      // Create a standalone network when no existing IDs are provided.
-      this.vpc = new ec2.Vpc(this, "Vpc", {
-        vpcName: `cprosite-${props.tier}`,
-        maxAzs: 2,
-        natGateways: 1,
-        subnetConfiguration: [
-          {
-            cidrMask: 24,
-            name: "Public",
-            subnetType: ec2.SubnetType.PUBLIC,
-          },
-          {
-            cidrMask: 24,
-            name: "Private",
-            subnetType: ec2.SubnetType.PRIVATE_WITH_EGRESS,
-          },
-        ],
-      });
-
-      this.albSecurityGroup = new ec2.SecurityGroup(this, "AlbSecurityGroup", {
-        vpc: this.vpc,
-        securityGroupName: `cprosite-${props.tier}-alb`,
-        description: "Allow HTTP/HTTPS inbound to ALB",
-        allowAllOutbound: true,
-      });
-      this.albSecurityGroup.addIngressRule(
-        ec2.Peer.anyIpv4(),
-        ec2.Port.tcp(80),
-        "HTTP from internet"
-      );
-      this.albSecurityGroup.addIngressRule(
-        ec2.Peer.anyIpv4(),
-        ec2.Port.tcp(443),
-        "HTTPS from internet"
-      );
-
-      this.ecsSecurityGroup = new ec2.SecurityGroup(this, "EcsSecurityGroup", {
-        vpc: this.vpc,
-        securityGroupName: `cprosite-${props.tier}-ecs`,
-        description: "Allow traffic from ALB to ECS tasks",
-        allowAllOutbound: true,
-      });
-      this.ecsSecurityGroup.addIngressRule(
-        this.albSecurityGroup,
-        ec2.Port.tcp(80),
-        "Frontend port from ALB"
-      );
-    }
-
-    new cdk.CfnOutput(this, "VpcId", {
-      value: this.vpc.vpcId,
-      exportName: `cprosite-${props.tier}-vpc-id`,
-    });
-  }
-}
-
-interface StorageStackProps extends cdk.StackProps {
-  env?: cdk.Environment;
-  tier: string;
-  vpc: ec2.IVpc;
-  ecsSecurityGroup: ec2.ISecurityGroup;
-}
-
-// Compatibility stack: provisions EFS and an access point for SQLite persistence.
-export class StorageStack extends cdk.Stack {
-  public readonly fileSystem: efs.FileSystem;
-  public readonly accessPoint: efs.AccessPoint;
-
-  constructor(scope: Construct, id: string, props: StorageStackProps) {
-    super(scope, id, props);
-
-    const efsSecurityGroup = new ec2.SecurityGroup(this, "EfsSecurityGroup", {
-      vpc: props.vpc,
-      securityGroupName: `cprosite-${props.tier}-efs`,
-      description: "Allow NFS from ECS tasks",
-      allowAllOutbound: false,
-    });
-    efsSecurityGroup.addIngressRule(
-      props.ecsSecurityGroup,
-      ec2.Port.tcp(2049),
-      "NFS from ECS tasks"
-    );
-
-    this.fileSystem = new efs.FileSystem(this, "Database", {
-      vpc: props.vpc,
-      fileSystemName: `cprosite-db-${props.tier}`,
-      securityGroup: efsSecurityGroup,
-      encrypted: true,
-      removalPolicy: cdk.RemovalPolicy.RETAIN,
-      performanceMode: efs.PerformanceMode.GENERAL_PURPOSE,
-      throughputMode: efs.ThroughputMode.BURSTING,
-      vpcSubnets: { subnetType: ec2.SubnetType.PRIVATE_WITH_EGRESS },
-    });
-
-    this.accessPoint = this.fileSystem.addAccessPoint("DataAccessPoint", {
-      path: "/data",
-      createAcl: {
-        ownerGid: "1001",
-        ownerUid: "1001",
-        permissions: "750",
-      },
-      posixUser: {
-        gid: "1001",
-        uid: "1001",
-      },
-    });
-
-    new cdk.CfnOutput(this, "FileSystemId", {
-      value: this.fileSystem.fileSystemId,
-      exportName: `cprosite-${props.tier}-efs-id`,
-    });
-
-    new cdk.CfnOutput(this, "AccessPointId", {
-      value: this.accessPoint.accessPointId,
-      exportName: `cprosite-${props.tier}-efs-ap-id`,
-    });
-  }
-}
-
-interface EcsStackProps extends cdk.StackProps {
-  env?: cdk.Environment;
-  tier: string;
-  vpc: ec2.IVpc;
-  albSecurityGroup: ec2.ISecurityGroup;
-  ecsSecurityGroup: ec2.ISecurityGroup;
-  efsFileSystemId: string;
-  efsAccessPointId: string;
-}
-
-// Compatibility stack: full ECS/ALB deployment path for earlier split-stack workflows.
-export class EcsStack extends cdk.Stack {
-  constructor(scope: Construct, id: string, props: EcsStackProps) {
-    super(scope, id, props);
-
-    // Single shared ECR repository; backend/frontend are separated by tags.
-    const appRepo = ecr.Repository.fromRepositoryName(this, "AppRepo", "cprosite");
-
-    const cluster = new ecs.Cluster(this, "Cluster", {
-      vpc: props.vpc,
-      clusterName: `cprosite-${props.tier}`,
-      containerInsights: true,
-    });
-
-    const executionRole = new iam.Role(this, "TaskExecutionRole", {
-      roleName: `cprosite-${props.tier}-ecs-execution`,
-      assumedBy: new iam.ServicePrincipal("ecs-tasks.amazonaws.com"),
-      managedPolicies: [
-        iam.ManagedPolicy.fromAwsManagedPolicyName(
-          "service-role/AmazonECSTaskExecutionRolePolicy"
-        ),
-      ],
-    });
-
-    const taskRole = new iam.Role(this, "TaskRole", {
-      roleName: `cprosite-${props.tier}-ecs-task`,
-      assumedBy: new iam.ServicePrincipal("ecs-tasks.amazonaws.com"),
-    });
-
-    const fileSystemArn = cdk.Arn.format(
-      {
-        service: "elasticfilesystem",
-        resource: "file-system",
-        resourceName: props.efsFileSystemId,
-        arnFormat: cdk.ArnFormat.SLASH_RESOURCE_NAME,
-      },
-      this
-    );
-    const accessPointArn = cdk.Arn.format(
-      {
-        service: "elasticfilesystem",
-        resource: "access-point",
-        resourceName: props.efsAccessPointId,
-        arnFormat: cdk.ArnFormat.SLASH_RESOURCE_NAME,
-      },
-      this
-    );
-
-    taskRole.addToPolicy(
-      new iam.PolicyStatement({
-        actions: [
-          "elasticfilesystem:ClientMount",
-          "elasticfilesystem:ClientWrite",
-          "elasticfilesystem:ClientRootAccess",
-        ],
-        resources: [fileSystemArn],
-        conditions: {
-          StringEquals: {
-            "elasticfilesystem:AccessPointArn": accessPointArn,
-          },
-        },
-      })
-    );
-
-    const logGroup = new logs.LogGroup(this, "LogGroup", {
-      logGroupName: `/ecs/cprosite-${props.tier}`,
-      retention: logs.RetentionDays.ONE_MONTH,
-      removalPolicy: cdk.RemovalPolicy.DESTROY,
-    });
-
-    const taskDef = new ecs.FargateTaskDefinition(this, "TaskDef", {
-      family: `cprosite-${props.tier}`,
-      cpu: 1024,
-      memoryLimitMiB: 2048,
-      executionRole,
-      taskRole,
-      volumes: [
-        {
-          name: "database",
-          efsVolumeConfiguration: {
-            fileSystemId: props.efsFileSystemId,
-            transitEncryption: "ENABLED",
-            authorizationConfig: {
-              accessPointId: props.efsAccessPointId,
-              iam: "ENABLED",
-            },
-          },
-        },
-      ],
-    });
-
-    const backendContainer = taskDef.addContainer("backend", {
-      containerName: "backend",
-      image: ecs.ContainerImage.fromEcrRepository(
-        appRepo,
-        `backend-${props.tier}-latest`
-      ),
-      environment: {
-        NODE_ENV: "production",
-        DATABASE: "/deploy/database/cprosite.db",
-      },
-      portMappings: [{ containerPort: 10000, protocol: ecs.Protocol.TCP }],
-      logging: ecs.LogDriver.awsLogs({
-        logGroup,
-        streamPrefix: "backend",
-      }),
-      healthCheck: {
-        command: [
-          "CMD-SHELL",
-          "curl -sf http://localhost:10000/api/ping || exit 1",
-        ],
-        interval: cdk.Duration.seconds(30),
-        timeout: cdk.Duration.seconds(5),
-        retries: 3,
-        startPeriod: cdk.Duration.seconds(60),
-      },
-    });
-
-    backendContainer.addMountPoints({
-      containerPath: "/deploy/database",
-      sourceVolume: "database",
-      readOnly: false,
-    });
-
-    taskDef.addContainer("frontend", {
-      containerName: "frontend",
-      image: ecs.ContainerImage.fromEcrRepository(
-        appRepo,
-        `frontend-${props.tier}-latest`
-      ),
-      environment: {
-        API_HOST: "http://localhost:10000",
-      },
-      portMappings: [{ containerPort: 80, protocol: ecs.Protocol.TCP }],
-      logging: ecs.LogDriver.awsLogs({
-        logGroup,
-        streamPrefix: "frontend",
-      }),
-      dependsOn: [
-        {
-          containerName: "backend",
-          condition: ecs.ContainerDependencyCondition.HEALTHY,
-        },
-      ],
-    });
-
-    const alb = new elbv2.ApplicationLoadBalancer(this, "Alb", {
-      vpc: props.vpc,
-      internetFacing: true,
-      securityGroup: props.albSecurityGroup,
-      loadBalancerName: `cprosite-${props.tier}`,
-      vpcSubnets: { subnetType: ec2.SubnetType.PUBLIC },
-    });
-
-    const targetGroup = new elbv2.ApplicationTargetGroup(this, "TargetGroup", {
-      vpc: props.vpc,
-      targetGroupName: `cprosite-${props.tier}`,
-      port: 80,
-      protocol: elbv2.ApplicationProtocol.HTTP,
-      targetType: elbv2.TargetType.IP,
-      healthCheck: {
-        path: "/",
-        interval: cdk.Duration.seconds(30),
-        healthyThresholdCount: 2,
-        unhealthyThresholdCount: 3,
-        healthyHttpCodes: "200-399",
-      },
-    });
-
-    alb.addListener("HttpListener", {
-      port: 80,
-      defaultTargetGroups: [targetGroup],
-    });
-
-    const service = new ecs.FargateService(this, "Service", {
-      cluster,
-      taskDefinition: taskDef,
-      serviceName: `cprosite-${props.tier}`,
-      desiredCount: 1,
-      securityGroups: [props.ecsSecurityGroup],
-      vpcSubnets: { subnetType: ec2.SubnetType.PRIVATE_WITH_EGRESS },
-      platformVersion: ecs.FargatePlatformVersion.VERSION1_4,
-      enableECSManagedTags: true,
-      propagateTags: ecs.PropagatedTagSource.SERVICE,
-    });
-
-    service.attachToApplicationTargetGroup(targetGroup);
-
-    new cdk.CfnOutput(this, "LoadBalancerDns", {
-      value: alb.loadBalancerDnsName,
-      description: "Application Load Balancer DNS name",
-      exportName: `cprosite-${props.tier}-alb-dns`,
-    });
-
-    new cdk.CfnOutput(this, "BackendRepoUri", {
-      value: appRepo.repositoryUri,
-      exportName: `cprosite-${props.tier}-backend-repo`,
-    });
-
-    new cdk.CfnOutput(this, "FrontendRepoUri", {
-      value: appRepo.repositoryUri,
-      exportName: `cprosite-${props.tier}-frontend-repo`,
-    });
-
-    new cdk.CfnOutput(this, "ClusterName", {
-      value: cluster.clusterName,
-      exportName: `cprosite-${props.tier}-cluster`,
-    });
-
-    new cdk.CfnOutput(this, "ServiceName", {
-      value: service.serviceName,
-      exportName: `cprosite-${props.tier}-service`,
-    });
   }
 }
